+++ packages/composites/market-closure/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/market-closure-adapter
 
+## 1.3.58
+
+### Patch Changes
+
+- Updated dependencies [[`e780274`](https://github.com/smartcontractkit/external-adapters-js/commit/e7802744e399b95c3053daf9979da3c2254dd3a8)]:
+  - @chainlink/finnhub-adapter@3.2.0
+
 ## 1.3.57
 
 ### Patch Changes