+++ packages/sources/finnhub-secondary/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/finnhub-secondary-adapter
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [[`e780274`](https://github.com/smartcontractkit/external-adapters-js/commit/e7802744e399b95c3053daf9979da3c2254dd3a8)]:
+  - @chainlink/finnhub-adapter@3.2.0
+
 ## 0.5.1
 
 ### Patch Changes